chore(scripting): rename testing.ts to schemaCoverage.ts, remove consumer test
