+++ src/scripting/schemaCoverage.ts
